requirements.txt
-e .
numpy==2.1.1
scipy==1.14.1
pandas==2.2.2
matplotlib==3.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.2
mlflow==2.16.1
fastapi==0.115.0
bcrypt==4.2.0
pyjwt==2.9.0
uvicorn==0.30.6
httpx==0.27.2
gdown==5.2.0
email-validator==2.2.0
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.4.1
torchvision==0.19.1
torchmetrics==1.4.2
torchinfo==1.8.0 
